move BatchedCubicSumcheck trait into sumcheck.rs
